code refactoring, getting rid of wide aws UserType usage
