+++ src/main/java/gov/ca/cwds/idm/service/cognito/CognitoServiceFacadeImpl.java
@@ -231,9 +231,7 @@ public class CognitoServiceFacadeImpl implements CognitoServiceFacade {
    * {@inheritDoc}
    */
   @Override
-  public void changeUserEnabledStatus(User existedUser,  Boolean newValue) {
-
-    String id = existedUser.getId();
+  public void changeUserEnabledStatus(String id,  Boolean newValue) {
     if (newValue) {
       AdminEnableUserRequest adminEnableUserRequest =
           new AdminEnableUserRequest().withUsername(id).withUserPoolId(properties.getUserpool());
