--- src/main/java/gov/ca/cwds/idm/service/cognito/attribute/AttributeHelper.java
@@ -1,11 +1,0 @@
-package gov.ca.cwds.idm.service.cognito.attribute;
-
-import com.amazonaws.services.cognitoidp.model.AttributeType;
-import com.amazonaws.services.cognitoidp.model.UserType;
-
-public interface AttributeHelper<T> {
-
-  T getAttributeValue(UserType userType, UserAttribute userAttribute);
-
-  AttributeType createAttribute(UserAttribute userAttribute, T value);
-}
